rename: step1, step2로 파일들 이동
