+++ src/main/java/step2/CalcOperator.java
@@ -1,4 +1,4 @@
-package calculator;
+package step2;
 
 import java.util.Arrays;
 import java.util.function.BiFunction;
