+++ src/main/java/step2/Calculator.java
@@ -1,4 +1,4 @@
-package calculator;
+package step2;
 
 public class Calculator {
