+++ src/test/java/step1/SetTest.java
@@ -1,3 +1,5 @@
+package step1;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
