+++ src/test/java/step2/CalcOperatorTest.java
@@ -1,4 +1,4 @@
-package calculator;
+package step2;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
