+++ src/test/java/step2/CalculatorTest.java
@@ -1,4 +1,4 @@
-package calculator;
+package step2;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
